pyproject.toml
[project]
name = "adk_agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Dr. Michael Menzel",email = "mugglmenzel@gmail.com"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=1.6.1,<2.0.0)",
    "google-genai (>=1.25.0,<2.0.0)",
    "google-cloud-aiplatform[adk,agent-engines] (>=1.102.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
